+++ src/udp.rs
@@ -8,36 +8,24 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::thread;
 
-
 use bytes::{BufMut, BytesMut};
 use futures::future::empty;
 use futures::sync::mpsc::Sender;
 use futures::IntoFuture;
 use net2::unix::UnixUdpBuilderExt;
 use net2::UdpBuilder;
-use slog::{Logger, info, warn, debug, o};
+use slog::{debug, info, o, warn, Logger};
 use std::os::unix::io::AsRawFd;
 use tokio::net::UdpSocket;
-use tokio::runtime::current_thread::Runtime;
 use tokio::reactor::Handle;
+use tokio::runtime::current_thread::Runtime;
 
 use crate::config::System;
 use crate::server::StatsdServer;
 use crate::task::Task;
 use crate::{DROPS, INGRESS};
 
-pub(crate) fn start_sync_udp(
-    log: Logger,
-    listen: SocketAddr,
-    chans: &Vec<Sender<Task>>,
-    config: Arc<System>,
-    n_threads: usize,
-    bufsize: usize,
-    mm_packets: usize,
-    mm_async: bool,
-    mm_timeout: u64,
-    flush_flags: Arc<Vec<AtomicBool>>,
-    ) {
+pub(crate) fn start_sync_udp(log: Logger, listen: SocketAddr, chans: &Vec<Sender<Task>>, config: Arc<System>, n_threads: usize, bufsize: usize, mm_packets: usize, mm_async: bool, mm_timeout: u64, flush_flags: Arc<Vec<AtomicBool>>) {
     info!(log, "multimessage enabled, starting in sync UDP mode"; "socket-is-blocking"=>!mm_async, "packets"=>mm_packets);
 
     // It is crucial for recvmmsg to have one socket per many threads
@@ -49,11 +37,7 @@ pub(crate) fn start_sync_udp(
     let sck = socket.bind(listen).unwrap();
     sck.set_nonblocking(mm_async).unwrap();
 
-    let mm_timeout = if mm_timeout == 0 {
-        config.network.buffer_flush_time
-    } else {
-        mm_timeout
-    };
+    let mm_timeout = if mm_timeout == 0 { config.network.buffer_flush_time } else { mm_timeout };
 
     for i in 0..n_threads {
         let chans = chans.clone();
@@ -131,11 +115,7 @@ pub(crate) fn start_sync_udp(
                     let mut chunks = Vec::with_capacity(mm_packets);
 
                     for i in 0..mm_packets {
-                        let chunk = iovec {
-                            iov_base: recv_buffer[i * rowsize..i * rowsize + rowsize].as_mut_ptr()
-                                as *mut c_void,
-                                iov_len: rowsize,
-                        };
+                        let chunk = iovec { iov_base: recv_buffer[i * rowsize..i * rowsize + rowsize].as_mut_ptr() as *mut c_void, iov_len: rowsize };
                         chunks.push(chunk);
                         // put the result to mheaders
                         mheaders[i].msg_hdr.msg_iov = &mut chunks[i];
@@ -146,31 +126,9 @@ pub(crate) fn start_sync_udp(
                         debug!(log, "recvmsg start");
                         // timeout is mutable and changed by every recvmmsg call, so it MUST be inside loop
                         // creating timeout as &mut fails because it's supposedly not dropped
-                        let mut timeout = if mm_timeout > 0 {
-                            timespec {
-                                tv_sec: (mm_timeout / 1000u64) as i64,
-                                tv_nsec: ((mm_timeout % 1000u64) * 1_000_000u64) as i64,
-                            }
-                        } else {
-                            timespec {
-                                tv_sec: 0,
-                                tv_nsec: 0,
-                            }
-                        };
+                        let mut timeout = if mm_timeout > 0 { timespec { tv_sec: (mm_timeout / 1000u64) as i64, tv_nsec: ((mm_timeout % 1000u64) * 1_000_000u64) as i64 } } else { timespec { tv_sec: 0, tv_nsec: 0 } };
 
-                        let res = unsafe {
-                            recvmmsg(
-                                fd as c_int,
-                                mhptr,
-                                mhlen as c_uint,
-                                flags,
-                                if mm_timeout > 0 {
-                                    &mut timeout
-                                } else {
-                                    null_mut()
-                                },
-                                )
-                        };
+                        let res = unsafe { recvmmsg(fd as c_int, mhptr, mhlen as c_uint, flags, if mm_timeout > 0 { &mut timeout } else { null_mut() }) };
 
                         if res == 0 {
                             // skip this shit
@@ -184,9 +142,7 @@ pub(crate) fn start_sync_udp(
                                 INGRESS.fetch_add(mlen, Ordering::Relaxed);
 
                                 // create address entry in messagemap
-                                let entry = bufmap
-                                    .entry(addrs[i])
-                                    .or_insert(BytesMut::with_capacity(mlen));
+                                let entry = bufmap.entry(addrs[i]).or_insert(BytesMut::with_capacity(mlen));
 
                                 // check we can fit the buffer
                                 if entry.remaining_mut() < mlen + 1 {
@@ -196,7 +152,7 @@ pub(crate) fn start_sync_udp(
                                 // and put the buffer into the map
                                 entry.put(&recv_buffer[i * rowsize..i * rowsize + mlen]);
 
-                                // reset addres to be used in next cycle
+                                // reset address to be used in next cycle
                                 addrs[i] = [0; 20];
                                 mheaders[i].msg_hdr.msg_namelen = 20;
                             }
@@ -222,20 +178,15 @@ pub(crate) fn start_sync_udp(
                                         let mut hasher = DefaultHasher::new();
                                         hasher.write(&addr);
                                         let ahash = hasher.finish();
-                                        let mut chan = if config.metrics.consistent_parsing {
-                                            chans[ahash as usize % chlen].clone()
-                                        } else {
-                                            ichans.next().unwrap().clone()
-                                        };
+                                        let mut chan = if config.metrics.consistent_parsing { chans[ahash as usize % chlen].clone() } else { ichans.next().unwrap().clone() };
                                         chan.try_send(Task::Parse(ahash, buf.take()))
                                             .map_err(|_| {
                                                 warn!(log, "error sending buffer(queue full?)");
-                                                DROPS.fetch_add(
-                                                    messages as usize,
-                                                    Ordering::Relaxed,
-                                                    );
-                                            }).unwrap_or(());
-                                    }).last();
+                                                DROPS.fetch_add(messages as usize, Ordering::Relaxed);
+                                            })
+                                            .unwrap_or(());
+                                    })
+                                    .last();
                             }
                         } else {
                             let errno = unsafe { *__errno_location() };
@@ -249,21 +200,12 @@ pub(crate) fn start_sync_udp(
                         }
                     }
                 }
-            }).expect("starting multimsg thread");
+            })
+            .expect("starting multimsg thread");
     }
 }
 
-pub(crate) fn start_async_udp(
-    log: Logger,
-    listen: SocketAddr,
-    chans: &Vec<Sender<Task>>,
-    config: Arc<System>,
-    n_threads: usize,
-    greens: usize,
-    async_sockets: usize,
-    bufsize: usize,
-    flush_flags: Arc<Vec<AtomicBool>>,
-    ) {
+pub(crate) fn start_async_udp(log: Logger, listen: SocketAddr, chans: &Vec<Sender<Task>>, config: Arc<System>, n_threads: usize, greens: usize, async_sockets: usize, bufsize: usize, flush_flags: Arc<Vec<AtomicBool>>) {
     info!(log, "multimessage is disabled, starting in async UDP mode");
 
     // Create a pool of listener sockets
@@ -278,10 +220,7 @@ pub(crate) fn start_async_udp(
 
     for i in 0..n_threads {
         // Each thread gets the clone of a socket pool
-        let sockets = sockets
-            .iter()
-            .map(|s| s.try_clone().unwrap())
-            .collect::<Vec<_>>();
+        let sockets = sockets.iter().map(|s| s.try_clone().unwrap()).collect::<Vec<_>>();
 
         let chans = chans.clone();
         let flush_flags = flush_flags.clone();
@@ -301,30 +240,16 @@ pub(crate) fn start_async_udp(
                         let chans = chans.clone();
                         // create UDP listener
                         let socket = socket.try_clone().expect("cloning socket");
-                        let socket =
-                            UdpSocket::from_std(socket, &Handle::default())
-                            .expect("adding socket to event loop");
+                        let socket = UdpSocket::from_std(socket, &Handle::default()).expect("adding socket to event loop");
 
-                        let server = StatsdServer::new(
-                            socket,
-                            chans.clone(),
-                            HashMap::new(),
-                            config.clone(),
-                            bufsize,
-                            0,
-                            i,
-                            readbuf,
-                            flush_flags.clone(),
-                            i,
-                            );
+                        let server = StatsdServer::new(socket, chans.clone(), HashMap::new(), config.clone(), bufsize, 0, i, readbuf, flush_flags.clone(), i);
 
                         runtime.spawn(server.into_future());
                     }
                 }
 
-                runtime
-                    .block_on(empty::<(), ()>())
-                    .expect("starting runtime for async UDP");
-            }).expect("creating UDP reader thread");
+                runtime.block_on(empty::<(), ()>()).expect("starting runtime for async UDP");
+            })
+            .expect("creating UDP reader thread");
     }
 }